Created visualizations
Didn't resize the dots in the scatter plots, we can try to do that tomorrow in tutorial. Also do we want to change figure 2 so that we have the other two variables compared against age so it's more uniform? It wouldn't be too hard to do and it might make it look neater since we have an odd number of variables and can't do pairs.